# openapi.yaml
openapi: 3.1.0
info:
  title: API INBCM
  description: ""
  contact:
    name: Support
    email: nocsinbcmdev@gmail.com
  version: 1.0.0
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
security:
  - {}
  - basicAuth: []
paths:
  /api/criarMuseu:
    post:
      summary: Cria um novo museu.
      description: Endpoint para criar um novo museu.
      tags:
        - Museu
      parameters:
        - in: body
          name: museu
          description: Dados do museu a ser criado.
          required: true
          schema:
            type: object
            properties:
              nome:
                type: string
                description: Nome do museu.
              endereco:
                type: object
                properties:
                  cidade:
                    type: string
                    description: Nome da cidade onde o museu está localizado.
                  UF:
                    type: string
                    enum:
                      - AC
                      - AL
                      - AP
                      - AM
                      - BA
                      - CE
                      - DF
                      - ES
                      - GO
                      - MA
                      - MT
                      - MS
                      - MG
                      - PA
                      - PB
                      - PR
                      - PE
                      - PI
                      - RJ
                      - RN
                      - RS
                      - RO
                      - RR
                      - SC
                      - SP
                      - SE
                      - TO
                    description: Sigla do estado onde o museu está localizado.
                  logradouro:
                    type: string
                    description: Nome da rua ou logradouro onde o museu está localizado.
                  numero:
                    type: string
                    description: Número do endereço onde o museu está localizado.
                  complemento:
                    type: string
                    description: Complemento do endereço (opcional).
                  bairro:
                    type: string
                    description: Bairro onde o museu está localizado.
                  cep:
                    type: string
                    description: CEP do endereço onde o museu está localizado.
                  municipio:
                    type: string
                    description: Nome do município onde o museu está localizado.
                  uf:
                    type: string
                    description: Sigla do estado onde o museu está localizado.
      responses:
        "200":
          description: Museu criado com sucesso.
        "400":
          description: Erro ao criar o museu.
  /api/listarMuseus:
    get:
      summary: Lista os museus.
      description: Endpoint para listar todos os museus cadastrados no sistema.
      tags:
        - Museu
      responses:
        "200":
          description: Lista de museus retornada com sucesso.
          content:
            application/json:
              schema:
                type: array
                items: {}
        "500":
          description: Erro ao listar museus.
  /api/museus:
    get:
      summary: Lista os museus do usuário.
      description: Endpoint para listar os museus associados ao usuário autenticado.
      tags:
        - Museu
      responses:
        "200":
          description: Lista de museus do usuário retornada com sucesso.
          content:
            application/json:
              schema:
                type: array
                items: {}
        "500":
          description: Erro ao listar museus do usuário.
  /api/uploads/{museu}/{anoDeclaracao}:
    post:
      summary: Envia uma declaração para o museu.
      description: Endpoint para enviar uma declaração para o museu especificado.
      tags:
        - Declarações
      parameters:
        - in: path
          name: museu
          description: ID do museu para o qual a declaração está sendo enviada.
          required: true
          schema:
            type: string
        - in: path
          name: anoDeclaracao
          description: Ano da declaração.
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                arquivisticoArquivo:
                  type: string
                  format: binary
                  description: Arquivo arquivístico.
                bibliograficoArquivo:
                  type: string
                  format: binary
                  description: Arquivo bibliográfico.
                museologicoArquivo:
                  type: string
                  format: binary
                  description: Arquivo museológico.
                arquivistico:
                  type: string
                  description: Dados arquivísticos.
                bibliografico:
                  type: string
                  description: Dados bibliográficos.
                museologico:
                  type: string
                  description: Dados museológicos.
                arquivisticoErros:
                  type: string
                  description: Erros no envio dos dados arquivísticos.
                bibliograficoErros:
                  type: string
                  description: Erros no envio dos dados bibliográficos.
                museologicoErros:
                  type: string
                  description: Erros no envio dos dados museológicos.
      responses:
        "200":
          description: Declaração enviada com sucesso.
        "400":
          description: Museu inválido.
        "500":
          description: Erro ao enviar arquivos para a declaração.
  /api/retificar/{museu}/{anoDeclaracao}/{idDeclaracao}:
    put:
      summary: Retifica uma declaração existente.
      description: Endpoint para retificar uma declaração existente para o museu
        especificado.
      tags:
        - Declarações
      parameters:
        - in: path
          name: museu
          description: ID do museu ao qual a declaração pertence.
          required: true
          schema:
            type: string
        - in: path
          name: anoDeclaracao
          description: Ano da declaração a ser retificada.
          required: true
          schema:
            type: string
        - in: path
          name: idDeclaracao
          description: ID da declaração a ser retificada.
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                arquivisticoArquivo:
                  type: string
                  format: binary
                  description: Novo arquivo arquivístico para a retificação.
                bibliograficoArquivo:
                  type: string
                  format: binary
                  description: Novo arquivo bibliográfico para a retificação.
                museologicoArquivo:
                  type: string
                  format: binary
                  description: Novo arquivo museológico para a retificação.
                arquivistico:
                  type: string
                  description: Novos dados arquivísticos para a retificação.
                bibliografico:
                  type: string
                  description: Novos dados bibliográficos para a retificação.
                museologico:
                  type: string
                  description: Novos dados museológicos para a retificação.
                arquivisticoErros:
                  type: string
                  description: Novos erros no envio dos dados arquivísticos para a retificação.
                bibliograficoErros:
                  type: string
                  description: Novos erros no envio dos dados bibliográficos para a retificação.
                museologicoErros:
                  type: string
                  description: Novos erros no envio dos dados museológicos para a retificação.
      responses:
        "200":
          description: Declaração retificada com sucesso.
        "404":
          description: Declaração não encontrada para o ano especificado.
        "500":
          description: Erro ao retificar declaração.
  /api/download/{museu}/{anoDeclaracao}/{tipoArquivo}:
    get:
      summary: Baixa um arquivo de declaração.
      description: Endpoint para baixar um arquivo de declaração para o museu e ano
        especificados.
      tags:
        - Declarações
      parameters:
        - in: path
          name: museu
          description: ID do museu ao qual a declaração pertence.
          required: true
          schema:
            type: string
        - in: path
          name: anoDeclaracao
          description: Ano da declaração do arquivo a ser baixado.
          required: true
          schema:
            type: string
        - in: path
          name: tipoArquivo
          description: Tipo de arquivo a ser baixado (arquivistico, bibliografico ou
            museologico).
          required: true
          schema:
            type: string
            enum:
              - arquivistico
              - bibliografico
              - museologico
      responses:
        "200":
          description: Arquivo de declaração baixado com sucesso.
        "404":
          description: Declaração não encontrada para o ano especificado ou arquivo não
            encontrado para o tipo especificado.
        "500":
          description: Erro ao baixar arquivo da declaração.
  /api/declaracoes:
    get:
      summary: Obtém todas as declarações do usuário.
      description: Endpoint para obter todas as declarações pertencentes ao usuário
        autenticado.
      tags:
        - Declarações
      responses:
        "200":
          description: Lista de todas as declarações do usuário obtida com sucesso.
        "500":
          description: Erro ao buscar declarações.
  /api/declaracoes/{id}:
    get:
      summary: Obtém uma declaração pelo ID.
      description: Endpoint para obter uma declaração específica pelo seu ID.
      tags:
        - Declarações
      parameters:
        - in: path
          name: id
          description: ID da declaração a ser obtida.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Declaração obtida com sucesso.
        "404":
          description: Declaração não encontrada para o ID especificado.
        "500":
          description: Erro ao buscar declaração.
  /api/declaracoes/{museu}/{anoDeclaracao}:
    get:
      summary: Obtém declarações por museu e ano.
      description: Endpoint para obter declarações de um museu específico para um ano
        específico.
      tags:
        - Declarações
      parameters:
        - in: path
          name: museu
          description: ID do museu.
          required: true
          schema:
            type: string
        - in: path
          name: anoDeclaracao
          description: Ano da declaração.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Declarações obtidas com sucesso.
        "404":
          description: Declarações não encontradas para o museu e ano especificados.
        "500":
          description: Erro ao buscar declarações.
  /api/declaracoesFiltradas:
    get:
      summary: Obtém declarações com base em filtros.
      description: Endpoint para buscar declarações com base em filtros especificados.
      tags:
        - Declarações
      responses:
        "200":
          description: Declarações filtradas obtidas com sucesso.
        "500":
          description: Erro ao buscar declarações com filtros.
    post:
      summary: Obtém declarações com base em filtros.
      description: Endpoint para buscar declarações com base em filtros especificados.
      tags:
        - Declarações
      responses:
        "200":
          description: Declarações filtradas obtidas com sucesso.
        "500":
          description: Erro ao buscar declarações com filtros.
...